+++ Runtime/RenderFeatures/VirtualMeshRenderFeature.cs
@@ -39,6 +39,7 @@ namespace Unity.VirtualMesh.Runtime
         private DrawShadowPass m_DrawShadowPass;
         private RefreshDepthPyramidPass m_RefreshDepthPass;
         private VisibilityPass m_VisibilityPass;
+        private DrawPass m_DrawPass;
 
         public override void Create()
         {
@@ -49,6 +50,7 @@ namespace Unity.VirtualMesh.Runtime
             m_DrawShadowPass = new DrawShadowPass(name + " Draw Shadow Pass");
             m_RefreshDepthPass = new RefreshDepthPyramidPass(name + " Refresh Depth Pyramid Pass");
             m_VisibilityPass = new VisibilityPass(name + " Visibility Pass");
+            m_DrawPass = new DrawPass(name + " Draw Pass");
         }
 
         protected override void Dispose(bool disposing)
@@ -82,6 +84,14 @@ namespace Unity.VirtualMesh.Runtime
             m_VisibilityPass.renderPassEvent = RenderPassEvent.BeforeRenderingGbuffer;
             m_VisibilityPass.settings = settings;
             renderer.EnqueuePass(m_VisibilityPass);
+
+            if (VirtualMeshManager.Instance.IsInternalDrawPassEnabled)
+            {
+                // drawing should occur after visibility has been resolved
+                m_DrawPass.renderPassEvent = RenderPassEvent.BeforeRenderingGbuffer;
+                m_DrawPass.settings = settings;
+                renderer.EnqueuePass(m_DrawPass);
+            }
         }
     }
 
@@ -549,6 +559,82 @@ namespace Unity.VirtualMesh.Runtime
                 passData.settings = settings;
                 passData.profilingSampler = m_ProfilingSampler;
                 passData.shadowCascadeCount = frameData.Get<UniversalShadowData>().mainLightShadowCascadesCount;
+
+                builder.AllowPassCulling(false);
+
+                builder.SetRenderFunc((PassData data, UnsafeGraphContext context) => ExecutePass(data, context));
+            }
+        }
+    }
+
+    /// <summary>
+    /// Custom render pass to draw newly disoccluded geometry (second pass in the two-pass occlusion culling algorithm).
+    /// </summary>
+    internal class DrawPass : ScriptableRenderPass
+    {
+        public VirtualMeshRenderFeature.Settings settings;
+
+        private string m_ProfilerTag;
+        private ProfilingSampler m_ProfilingSampler;
+
+        private class PassData
+        {
+            internal VirtualMeshRenderFeature.Settings settings;
+            internal ProfilingSampler profilingSampler;
+        }
+
+        public DrawPass(string tag)
+        {
+            m_ProfilerTag = tag;
+            m_ProfilingSampler = new ProfilingSampler(m_ProfilerTag);
+        }
+
+        static void ExecutePass(PassData data, UnsafeGraphContext context)
+        {
+            var cmd = context.cmd;
+            using (new ProfilingScope(cmd, data.profilingSampler))
+            {
+                // draw geometry
+                for (int i = 0; i < VirtualMeshManager.Instance.Materials.Length; i++)
+                {
+                    uint vertexCount = VirtualMeshManager.Instance.MaterialVertexCounts[i];
+                    if (i > 0 && vertexCount == 0)
+                        continue;
+
+                    var material = VirtualMeshManager.Instance.Materials[i];
+
+                    // try to find the correct pass (TODO improve this)
+                    var pass = material.FindPass(data.settings.renderMode == RenderingPath.DeferredShading ? VirtualMeshShaderProperties.GBufferPassName : VirtualMeshShaderProperties.SGForwardPassName);
+                    if (pass == -1)
+                        pass = material.FindPass(VirtualMeshShaderProperties.LitForwardPassName);
+
+                    if (pass != -1)
+                        cmd.DrawProceduralIndirect(VirtualMeshManager.Instance.CompactedIndexBuffer, Matrix4x4.identity, material, pass, MeshTopology.Triangles, VirtualMeshManager.Instance.DrawArgsBuffer, i * VirtualMeshManager.Instance.DrawArgsBuffer.stride);
+                }
+            }
+        }
+
+        public override void RecordRenderGraph(RenderGraph renderGraph, ContextContainer frameData)
+        {
+            if (VirtualMeshManager.Instance == null || !VirtualMeshManager.Instance.IsInitialized)
+                return;
+
+            var resourceData = frameData.Get<UniversalResourceData>();
+
+            using (var builder = renderGraph.AddUnsafePass<PassData>(m_ProfilerTag, out var passData))
+            {
+                passData.settings = settings;
+                passData.profilingSampler = m_ProfilingSampler;
+
+                if (settings.renderMode == RenderingPath.DeferredShading)
+                {
+                    for (int i = 0; i < resourceData.gBuffer.Length; i++)
+                        builder.SetRenderAttachment(resourceData.gBuffer[i], i, AccessFlags.Write);
+                }
+                else
+                    builder.SetRenderAttachment(resourceData.cameraColor, 0, AccessFlags.Write);
+
+                builder.SetRenderAttachmentDepth(resourceData.cameraDepth, AccessFlags.Write);
 
                 builder.AllowPassCulling(false);